requirements.txt
# Network scanning and analysis
scapy==2.5.0
netifaces==0.11.0
python-nmap==0.7.1

# Web framework and API
fastapi==0.95.1
uvicorn==0.22.0
jinja2==3.1.2
python-multipart==0.0.6

# Database
sqlalchemy==2.0.12
alembic==1.10.4
aiosqlite==0.19.0

# Utilities
pyyaml==6.0
python-dotenv==1.0.0
schedule==1.2.0
mac-vendor-lookup==0.1.12
requests==2.29.0

# Testing
pytest==7.3.1
pytest-asyncio==0.21.0
